gm: split all downstream logic into separate file
Since all CR watch actions are placed into corresponding controller,
controllers/downstream.go is unnecessary.
