+++ pkg/globalmanager/controllers/jointinference/downstream.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package dataset
+package jointinference
 
 import (
 	"fmt"
@@ -25,29 +25,23 @@ import (
 	"github.com/kubeedge/sedna/pkg/globalmanager/runtime"
 )
 
-// syncToEdge syncs the dataset resources
 func (c *Controller) syncToEdge(eventType watch.EventType, obj interface{}) error {
-	dataset, ok := obj.(*sednav1.Dataset)
+	joint, ok := obj.(*sednav1.JointInferenceService)
 	if !ok {
 		return nil
 	}
 
 	// Here only propagate to the nodes with non empty name
-	nodeName := dataset.Spec.NodeName
+	// FIXME: only the case that Spec.NodeName specified is support
+	nodeName := joint.Spec.EdgeWorker.Template.Spec.NodeName
 	if len(nodeName) == 0 {
 		return fmt.Errorf("empty node name")
 	}
 
-	// Since t.Kind may be empty,
-	// we need to fix the kind here if missing.
-	// more details at https://github.com/kubernetes/kubernetes/issues/3030
-	if len(dataset.Kind) == 0 {
-		dataset.Kind = KindName
+	if len(joint.Kind) == 0 {
+		joint.Kind = KindName
 	}
-
-	runtime.InjectSecretAnnotations(c.kubeClient, dataset, dataset.Spec.CredentialName)
-
-	return c.sendToEdgeFunc(nodeName, eventType, dataset)
+	return c.sendToEdgeFunc(nodeName, eventType, joint)
 }
 
 func (c *Controller) SetDownstreamSendFunc(f runtime.DownstreamSendFunc) error {
