+++ pkg/globalmanager/controllers/lifelonglearning/lifelonglearningjob.go
@@ -28,6 +28,7 @@ import (
 	utilrand "k8s.io/apimachinery/pkg/util/rand"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/apimachinery/pkg/util/wait"
+	"k8s.io/apimachinery/pkg/watch"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/scheme"
 	v1core "k8s.io/client-go/kubernetes/typed/core/v1"
@@ -80,6 +81,8 @@ type Controller struct {
 	recorder record.EventRecorder
 
 	cfg *config.ControllerConfig
+
+	sendToEdgeFunc runtime.DownstreamSendFunc
 }
 
 // Run starts the main goroutine responsible for watching and syncing jobs.
@@ -726,12 +729,15 @@ func New(cc *runtime.ControllerContext) (runtime.FeatureControllerI, error) {
 	jobInformer.Informer().AddEventHandler(cache.ResourceEventHandlerFuncs{
 		AddFunc: func(obj interface{}) {
 			jc.enqueueController(obj, true)
+			jc.syncToEdge(watch.Added, obj)
 		},
 		UpdateFunc: func(old, cur interface{}) {
 			jc.enqueueController(cur, true)
+			jc.syncToEdge(watch.Added, cur)
 		},
 		DeleteFunc: func(obj interface{}) {
 			jc.enqueueController(obj, true)
+			jc.syncToEdge(watch.Deleted, obj)
 		},
 	})
 	jc.jobLister = jobInformer.Lister()
